zeroclaw/savia-voice/requirements.txt
faster-whisper
silero-vad
sounddevice
numpy
edge-tts
websockets
pyyaml
